gencrud/requirements.txt
Django==2.2.10
# !django-suit==0.2.25
django-ckeditor==5.8.0
django-mptt==0.10.0
Pillow>=6.2.2
transliterate
gunicorn==20.0.0
pytz==2020.1
docutils==0.16
psycopg2==2.8.5
psycopg2-binary==2.8.5
easy-thumbnails==2.5
django-image-cropping==1.2.0
libsass==0.20.0
django-filebrowser==3.12.1
# django-filebrowser-no-grappelli==3.7.4
# sorl-thumbnail
django-simple-search==1.0.2
django-uuslug==1.1.9
django-htmlmin==0.11.0
django-daterange-filter==1.3.0
django-markdown-deux==1.0.5
django-pagedown==2.0.3
django-simple-math-captcha==1.0.9
django-compressor==2.3
django-import-export==2.0.2

djangorestframework==3.10.3
django-filter==2.2.0
markdown==3.2.1

